requirements.txt
python-dotenv
requests
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
pandas
python-telegram-bot
openai
